+++ components/features/landing/Landing.tsx
@@ -4,8 +4,8 @@ import { Features } from "./Features";
 import { HowItWorks } from "./HowItWorks";
 import { CTA } from "./CTA";
 import { Footer } from "./Footer";
-import { Navbar } from "@/components/shared/navbar";
-  
+import { Navbar } from "@/components/shared";
+   
 export function Landing() {
   return (
     <div className="min-h-screen bg-white">
